Worked out the task aliases, some order changes
To get it so that I could define task_aliases in the fabfile and still have
them picked up in time by main() I had to take the chunk in main() that found
then parsed the fabfile and put it at the top before the options were parsed.

This made it so that when refrencing the state.env.task_aliases that it was
actually there and populated in the env dict before the parse_options() was
called. As to why I wanted all this there instead of at parse_arguments()
that is called much later, it is so that flags can also be combined into the
task aliases in case they want things to be defined at 'run time' and not have
to make logic to define the options that fab flags offer.

All tests still pass, so I figure it's a go, especially since I added in the
tests for options, to cover more of exactly what I'm changing.